# ui/open-api-doc.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "Vacancy Management API",
        "description": "API для управления вакансиями",
        "contact": {
            "name": "Vacancy Management Team",
            "email": "support@vacancy.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://springdoc.org"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8081",
            "description": "Generated server url"
        }
    ],
...
    "paths": {
...
        "/api/offers/{id}": {
            "get": {
                "tags": [
                    "Offer Controller"
                ],
                "summary": "Получение оффера по ID",
                "operationId": "getOffer",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/OfferDto"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Offer Controller"
                ],
                "summary": "Обновление оффера",
                "operationId": "updateOffer",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/OfferDto"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/OfferDto"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Offer Controller"
                ],
                "summary": "Удаление оффера",
                "operationId": "deleteOffer",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/job-applications/{id}": {
            "get": {
                "tags": [
                    "Job Application Controller"
                ],
                "summary": "Получение отклика по ID",
                "operationId": "getJobApplication",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/JobApplicationDto"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Job Application Controller"
                ],
                "summary": "Обновление отклика",
                "operationId": "updateJobApplication",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/JobApplicationDto"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/JobApplicationDto"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Job Application Controller"
                ],
                "summary": "Удаление отклика",
                "operationId": "deleteJobApplication",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
        "/api/offers": {
            "get": {
                "tags": [
                    "Offer Controller"
                ],
                "summary": "Получение списка всех офферов",
                "operationId": "getAllOffers",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/OfferDto"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Offer Controller"
                ],
                "summary": "Создание нового оффера",
                "operationId": "createOffer",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/OfferDto"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/OfferDto"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/job-applications": {
            "get": {
                "tags": [
                    "Job Application Controller"
                ],
                "summary": "Получение списка всех откликов",
                "operationId": "getAllJobApplications",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/JobApplicationDto"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Job Application Controller"
                ],
                "summary": "Создание нового отклика",
                "operationId": "createJobApplication",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/JobApplicationDto"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/JobApplicationDto"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/auth/register": {
            "post": {
                "tags": [
                    "auth-controller"
                ],
                "operationId": "register",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserRegisterDto"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserDto"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/auth/login": {
            "post": {
                "tags": [
                    "auth-controller"
                ],
                "operationId": "login",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserLoginDto"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "OfferDto": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "description": "ID оффера",
                        "format": "int64",
                        "example": 1
                    },
                    "applicationId": {
                        "type": "integer",
                        "description": "ID заявки",
                        "format": "int64",
                        "example": 1
                    },
                    "offerText": {
                        "type": "string",
                        "description": "Текст оффера",
                        "example": "Предлагаем вам должность Java Developer"
                    },
                    "pdfFilePath": {
                        "type": "string",
                        "description": "Путь к PDF файлу оффера",
                        "example": "/offers/offer_1.pdf"
                    },
                    "createdAt": {
                        "type": "string",
                        "description": "Дата создания оффера",
                        "format": "date-time"
                    }
                },
                "description": "DTO для оффера"
            },
            "JobApplicationDto": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "description": "ID отклика",
                        "format": "int64",
                        "example": 1
                    },
                    "resumeId": {
                        "type": "integer",
                        "description": "ID резюме",
                        "format": "int64",
                        "example": 1
                    },
                    "vacancyId": {
                        "type": "integer",
                        "description": "ID вакансии",
                        "format": "int64",
                        "example": 1
                    },
                    "status": {
                        "type": "string",
                        "description": "Статус отклика",
                        "example": "new"
                    },
                    "matchScore": {
                        "type": "number",
                        "description": "Процент соответствия",
                        "format": "float",
                        "example": 85.5
                    },
                    "createdAt": {
                        "type": "string",
                        "description": "Дата создания отклика",
                        "format": "date-time"
                    }
                },
                "description": "DTO для отклика на вакансию"
            },
            "UserRegisterDto": {
                "type": "object",
                "properties": {
                    "username": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string"
                    }
                }
            },
            "UserDto": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "username": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "role": {
                        "type": "string"
                    }
                }
            },
            "UserLoginDto": {
                "type": "object",
                "properties": {
                    "email": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string"
                    }
                }
            }
        }
    }
}